Improve the responsiveness of disassembly view.
There has always been an issue that the second time user chooses a
function from the function table, it requires much longer to render the
CFG than the first time. After much debugging, it turns out that when
"conditional" is used, Block becomes a child of ui.CFGNode after
initialization, and the second time ui.CFGNode is created, both the old
Block instance and a new Block instance generated from the Conditional
will be activated and refreshed.

This is probably a bug in enaml. A temporary workaround is to not use
conditional in performance-critical part.