The first version rewritten on PySide.
From this commit, we no longer depend on enaml. Enaml is great, but
there are several fundamental issues (especially the speed of the layout
engine) makes it unfeasible to use by angr-management.

For now, we stay with PySide. It seems pretty OK, and the speed is so
much better (try the new disassembly graph view).