A hackish workaround for missing ins_addr on Function graphs.
During Function graph normalization, when a new node is created, we do
not re-lift the old node, and thus we do not have any value for
`ins_addr` or `stmt_idx` to set onto the edge. This commit can be
reverted once the issue in normalization is solved by relifting.